+++ copy.go
@@ -25,6 +25,7 @@ import (
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 	"golang.org/x/sync/semaphore"
 	"oras.land/oras-go/v2/content"
+	"oras.land/oras-go/v2/descriptor"
 	"oras.land/oras-go/v2/errdef"
 	"oras.land/oras-go/v2/internal/cas"
 	"oras.land/oras-go/v2/internal/docker"
@@ -370,7 +371,7 @@ func resolveRoot(ctx context.Context, src ReadOnlyTarget, srcRef string, proxy *
 	defer rc.Close()
 	// cache root if it is a non-leaf node
 	fetcher := content.FetcherFunc(func(ctx context.Context, target ocispec.Descriptor) (io.ReadCloser, error) {
-		if Equal(target, root) {
+		if descriptor.Equal(target, root) {
 			return rc, nil
 		}
 		return nil, errors.New("fetching only root node expected")
@@ -395,7 +396,7 @@ func prepareCopy(ctx context.Context, dst Target, dstRef string, proxy *cas.Prox
 					return err
 				}
 			}
-			if !Equal(desc, root) {
+			if !descriptor.Equal(desc, root) {
 				// for non-root node, do nothing
 				return nil
 			}
@@ -415,7 +416,7 @@ func prepareCopy(ctx context.Context, dst Target, dstRef string, proxy *cas.Prox
 	} else {
 		postCopy := opts.PostCopy
 		opts.PostCopy = func(ctx context.Context, desc ocispec.Descriptor) error {
-			if Equal(desc, root) {
+			if descriptor.Equal(desc, root) {
 				// for root node, tag it after copying it
 				if err := dst.Tag(ctx, root, dstRef); err != nil {
 					return err
@@ -435,7 +436,7 @@ func prepareCopy(ctx context.Context, dst Target, dstRef string, proxy *cas.Prox
 				return err
 			}
 		}
-		if !Equal(desc, root) {
+		if !descriptor.Equal(desc, root) {
 			return nil
 		}
 		// enforce tagging when root is skipped
