+++ descriptor/descriptor_test.go
@@ -1,4 +1,19 @@
-package oras_test
+/*
+Copyright The ORAS Authors.
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package descriptor
 
 import (
 	"reflect"
@@ -6,7 +21,6 @@ import (
 
 	"github.com/opencontainers/go-digest"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
-	"oras.land/oras-go/v2"
 )
 
 func TestGenerateDescriptor(t *testing.T) {
@@ -47,7 +61,7 @@ func TestGenerateDescriptor(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := oras.GenerateDescriptor(tt.args.content, tt.args.mediaType)
+			got, err := GenerateDescriptor(tt.args.content, tt.args.mediaType)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("GenerateDescriptor() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -131,7 +145,7 @@ func TestEqual(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if got := oras.Equal(tt.args.a, tt.args.b); got != tt.want {
+			if got := Equal(tt.args.a, tt.args.b); got != tt.want {
 				t.Errorf("Equal() = %v, want %v", got, tt.want)
 			}
 		})
